+++ magda-web-client/src/Components/Dataset/Add/Pages/ReviewPage/CollapseBoxContext.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const DEFAULT_OPEN_STATUS = false;
+
+const CollapseBoxContext = React.createContext(DEFAULT_OPEN_STATUS);
+
+export default CollapseBoxContext;